lbryd_api: correct class name `LbrydApi` in `lbryd_api.py`
In f540b8f2 the class was changed from `LbryApi` to `LbrydApi`.

However, the class was not changed in the `pybry/lbryd_api.py` file;
therefore, when running the `setup.py` it fails, as this setup imports
the entire `pybry` package to get the `__version__` variable.

Now it won't fail.