generator: update the parsing of the file as the json changed
The `generator.generate_lbryd_wrapper` method was last updated
in October 2018.
It takes a JSON file with all commands in the LBRY daemon,
`lbry/extras/daemon/daemon.py`.

This JSON file, `docs/api.json`, is normally retrieved
from the latest version in the `lbry-sdk` repository.

The `api.json` is produced by `scripts/generate_json_api.py`.
This script was updated at some point in 2019 to put the commands
in groups (0200756), which broke the `generate_lbryd_wrapper` method.

With this commit, we fix the issue so that now the latest version
of the `api.json` can be parsed.