generator: do not parse the `returns` key, it causes errors
The `generator.generate_method_definition` method
will extract the `'returns'` key from the command dictionary
and process it to find the return type, and return information.

However, since the `'returns'` value is not well standardized
at the moment (2021), it causes problems with the parsing,
and thus the output file `pybry/lbryd_api.py`
will not be correclty produced.

Maybe in the future we can re-enable this parsing once we can handle
the special cases.