docs: add two downloaded `api.json` files, obsolete and current
Version 2.0.0 of `pybry` is from September 2021 (fdac60338a).
The format of the JSON file that it requires has changed since 2018.
The included `docs/api.json` file
is from commit e1e76005 of the `lbry-sdk` repository.

Version 1.6.4 of `pybry` is from October 2018 (af86805a80).
The included `api.json-1.6.4` file
is from commit c753a582 of the `lbry-sdk` repository.

The newer `api.json` must be used with v2.0 of `generator.py`,
while `api.json-1.6.4` is the format expected by the obsolete v1.6.4 code,
and cannot be processed by v2.0 as the information is organized
in a different way.