docs: manually fix the downloaded `api.json` file
This particular JSON file has an error that needs to be
manually fixed so that the wrapped Python API is correctly written
by the `generator.py` module.

In the `file_list` command, the argument `blobs_in_stream`
has a wrong name.

The name is changed from
```
"name": "blobs_in_stream<blobs_in_stream>",
```

to
```
"name": "blobs_in_stream",
```

A `docs/README.md` is included to explain where the two `api.json`
files come from.

The newer `api.json` must be used with v2.0 of `generator.py`,
while `api.json-1.6.4` is the format expected by the obsolete v1.6.4 code,
and cannot be processed by v2.0 as the information is organized
in a different way.