pybry: release new version 2.1.0
The `generator.generate_lbryd_wrapper` now accepts as argument
the name of the offline JSON file to read.

Allow executing the `generator.py` module directly.

Include a proper `api.json` file so that we can create a working
API wrapper without an Internet connection.
This file comes from commit e1e76005 of the `lbry-sdk` repository.

Provide a `Makefile` to build the API wrapper, as it works better
than the existing `setup.py` file.

Update the `README.md` with the new build information.