template: move static files to `template/` directory
The `__init__.py` file is renamed to `_init.py`
so that the `template` directory does not try importing this module.
Once this is copied to the final library, it will be renamed `__init__.py`.

The `LBRYException.py` file is renamed to `exception.py`
as modules should be lowercase by convention. The exception
inside is still in `CamelCase`.