template: update the static modules
The renamed `exception.py` module is imported appropriately
in `_init.py` and `base_api.py`.

The code of `base_api.BaseApi` is organized and formatted.

The constants in `constants.py` are updated and made
more flexible, so that the values can be propagated better.

The `exception` module is made easier to read.
The class is renamed from `LBRYException` to `LBRYError`
to follow the convention that exceptions end with "Error".
The module and class docstrings are rewritten slightly,
and the code is organized a bit as well.