pybry: release new version 2.2.0
The last stable version is 2.1.0 (commit 4737c526ab).

The big change in this release is that the `pybry/` directory
is no longer part of the repository. This directory
is now created at build time when running `make` or `setup.py`.

* The files that are used during the building of the package
are now placed in the `template/` directory. Some files
are unmodified, and they are simply copied to the final `pybry/`
directory; others are used as templates to write
the final wrapper module.

* Two static files were renamed: `LBRYException.py` to `exception.py`,
and `__init__.py` to `_init.py`. We don't want to call it `__init__.py`
inside `template/` as this will cause import errors
when we try loading some values from this directory.
This `_init.py` file is renamed to `__init__.py`
once it is copied to the final `pybry/` directory.

* The static files are modified slightly, improving readability
and documentation.

* In `exception.py` the class is renamed from `LBRYException` to `LBRYError`
to follow the convention that exceptions end with "Error".

* In `constants.py` some constants are renamed,
and new ones are added to be more flexible.

* In `generator.py` the constants are now imported
from `template/` instead of `pybry/`.
This solves a problem that existed before, where the build process
would fail if another file had a syntax error.

* In `generator.py` a new function is added, `generate_basic_modules`
to create the `pybry/` directory if it doesn't exist,
and to copy the static modules from `template/` to `pybry/`.

* In `generator.py` a new function is added, `generate_lbrycrd_wrapper`
to generate the wrapper for the `lbrycrd` daemon.
Since there is no JSON file holding API information for `lbrycrd`,
this function basically just copies the template file
from `template/` to `pybry/`.

* In `generator.py` the `main` function is updated accordingly
to run `generate_basic_modules`, `generate_lbrycrd_wrapper`,
and `generate_lbryd_wrapper` in sequence.

* The `Makefile` is updated so that the `clean` target removes
the `pybry/` directory, which now can be recreated by running `make`.

* The `setup.py` is updated too so that the `build_local` and `build_online`
commands call the `generator.main` function with the appropriate arguments.
The `clean` command is also updated so that it removes the `pybry/` directory,
which now can be recreated by running the build functions.

* The `MANIFEST` and `README.md` files are updated.